Added Alma Redemptoris Mater
Unsure about whether I transcribed the ligatures correctly and about my
decisions with regard to the text underlay. What do we do when the text
of the example is at variance with the norms of spelling today (e.g., I
found “coeli” spelled “caeli”)? The example had what looked like a key
signature on each line so I decided to put a key signature in. With
regard to the time signature, 63 was the highest numerator possible so
I decided to make every line a bar and then hide the barlines. For ease
of transcription and to preserve line breaks, I made each bar
correspond to a line in the example. Perhaps this is something we may
want to do consistently?